+++ test/lmo/S1.cpp
@@ -4,7 +4,7 @@
 ///         S1(x, y) used in the Lagarias-Miller-Odlyzko
 ///         and Deleglise-Rivat prime counting algorithms.
 ///
-/// Copyright (C) 2023 Kim Walisch, <kim.walisch@gmail.com>
+/// Copyright (C) 2026 Kim Walisch, <kim.walisch@gmail.com>
 ///
 /// This file is distributed under the BSD License. See the COPYING
 /// file in the top level directory.
@@ -35,10 +35,10 @@ int main()
 {
   std::random_device rd;
   std::mt19937 gen(rd());
-  std::uniform_int_distribution<int> dist(0, 10000000);
+  std::uniform_int_distribution<int64_t> dist(0, 10000000);
   int threads = 1;
 
-  for (int i = 0; i < 1000; i++)
+  for (int64_t i = 0; i < 1000; i++)
   {
     int64_t x = dist(gen);
     int64_t y = iroot<3>(x);
